Refactoring for convention   * All requires of custom code are now aggregated in a single file   * There are too many files named sheldon - The CLI has been renamed to bin/launcher
