+++ bin/launcher
@@ -1,7 +1,5 @@
 #!/usr/bin/env ruby
-
 require "thor"
-require_relative "../lib/sheldon"
 
 class Launcher < Thor
